+++ code/src/types/backend-agent.ts
@@ -1,6 +1,12 @@
 import { EVENT_KEY } from "utils/constants";
 import { SparqlResponseField } from "./form";
 
+export const BackendApis = {
+  REGISTRY_BACKEND: process.env.REGISTRY_BACKEND_URL,
+  REGISTRY_TASK_ATTACHMENT: process.env.REGISTRY_TASK_ATTACHMENT_URL,
+  FILE_EXPORTER: process.env.FILE_EXPORTER_URL,
+};
+
 export type AgentResponseBody = {
   apiVersion: string;
   data?: {
@@ -14,6 +20,12 @@ export type AgentResponseBody = {
     message: string;
   };
 };
+
+export type FileResponse = {
+  blob: Blob;
+  file: string;
+};
+
 export type HistoryDetails = {
   message: SparqlResponseField;
   timestamp: SparqlResponseField;